+++ blobfuse2-nightly.yaml
@@ -36,6 +36,11 @@ parameters:
     type: boolean
     default: false
 
+  - name: scenario_tests
+    displayName: 'Run various scenario based workloads to check data integrity'
+    type: boolean
+    default: true
+
   - name: linux_git_test
     displayName: 'Compile Linux Source and Run git clone Tests'
     type: string
@@ -1150,6 +1155,82 @@ stages:
                 unmount: true
                 delete_containers: true
 
+  - ${{ if eq(parameters.scenario_tests, true) }}:
+    - stage: ScenarioTests
+      dependsOn: []
+      jobs:
+        # Ubuntu Tests
+        - job: Set_1
+          timeoutInMinutes: 60
+          strategy:
+            matrix:
+              Ubuntu-20:
+                AgentName: 'blobfuse-ubuntu20'
+                distro: 'ubuntu'
+                distro_version: 'ubuntu_20_x86'
+                fuselib: 'libfuse-dev'
+                tags: 'fuse2'
+              Ubuntu-22:
+                AgentName: 'blobfuse-ubuntu22'
+                distro: 'ubuntu'
+                distro_version: 'ubuntu_22_x86'
+                fuselib: 'libfuse3-dev'
+                tags: 'fuse3'
+
+          pool:
+            name: "blobfuse-ubuntu-pool"
+            demands:
+              - ImageOverride -equals $(AgentName)
+
+          variables:
+            - group: NightlyBlobFuse
+            - name: MOUNT_DIR
+              value: '$(Pipeline.Workspace)/blob_mnt'
+            - name: TEMP_DIR
+              value: '$(Pipeline.Workspace)/blobfuse2_tmp'
+            - name: BLOBFUSE2_CFG
+              value: '$(Pipeline.Workspace)/blobfuse2.yaml'
+            - name: BLOBFUSE2_ADLS_CFG
+              value: '$(Pipeline.Workspace)/blobfuse2.adls.yaml'
+            - name: skipComponentGovernanceDetection
+              value: true
+            - name: GOPATH
+              value: '$(Pipeline.Workspace)/go'
+            - name: ROOT_DIR
+              value: '$(Agent.TempDirectory)'
+            - name: WORK_DIR
+              value: '$(Build.SourcesDirectory)'
+
+          steps:
+            # -------------------------------------------------------
+            # Pull and build the code
+            - template: 'azure-pipeline-templates/build.yml'
+              parameters:
+                skip_ut: true
+
+            - template: 'azure-pipeline-templates/scenario.yml'
+              parameters:
+                cache_mode: file_cache
+                config_file: $(BLOBFUSE2_CFG)
+                account_name: $(BF2_BLK_ACC_NAME)
+                account_key: $(BF2_BLK_ACC_KEY)
+                account_type: block
+                verbose_log: ${{ parameters.verbose_log }}
+
+            # Commented out block cache scenario tests due to known issues
+            # - template: 'azure-pipeline-templates/scenario.yml'
+            #   parameters:
+            #     cache_mode: block_cache
+            #     config_file: $(BLOBFUSE2_CFG)
+            #     account_name: $(BF2_BLK_ACC_NAME)
+            #     account_key: $(BF2_BLK_ACC_KEY)
+            #     account_type: block
+            #     verbose_log: ${{ parameters.verbose_log }}
+
+            - template: 'azure-pipeline-templates/cleanup.yml'
+              parameters:
+                unmount: true
+                delete_containers: true
 
   - ${{ if ne(parameters.linux_git_test, 'none') }}:
     - stage: CompileLinux_GitClone
